Updated Gallery Templates
Updated the gallery templates as well as fixed a typo in the write
controller causing the edit action to load the add template.

Also wrote some javascript to style file inputs to make their appearance
match other bootstrap form elements as well as enable rendering a
preview of image files when selected.

The templates still need a bit of style work, but this gets them much
more inline with our other forms.